Move CredentialMakeResult to Fido2NetLib.Objects namespace
