+++ Src/Fido2/Fido2.cs
@@ -119,23 +119,6 @@ public class Fido2 : IFido2
 
         return result;
     }
-
-    /// <summary>
-    /// Result of parsing and verifying attestation. Used to transport Public Key back to RP
-    /// </summary>
-    public sealed class CredentialMakeResult : Fido2ResponseBase
-    {
-        public CredentialMakeResult(string status, string errorMessage, RegisteredPublicKeyCredential? result)
-        {
-            Status = status;
-            ErrorMessage = errorMessage;
-            Result = result;
-        }
-
-        public RegisteredPublicKeyCredential? Result { get; }
-
-        // todo: add debuginfo?
-    }
 }
 
 /// <summary>
