+++ Src/Fido2/IFido2.cs
@@ -22,7 +22,7 @@ public interface IFido2
         IsUserHandleOwnerOfCredentialIdAsync isUserHandleOwnerOfCredentialIdCallback,
         CancellationToken cancellationToken = default);
 
-    Task<Fido2.CredentialMakeResult> MakeNewCredentialAsync(
+    Task<CredentialMakeResult> MakeNewCredentialAsync(
         AuthenticatorAttestationRawResponse attestationResponse,
         CredentialCreateOptions origChallenge,
         IsCredentialIdUniqueToUserAsyncDelegate isCredentialIdUniqueToUser,
